LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-011-expected.txt


FAIL Should use the last fallback position initially assert_equals: Anchored element should be at the bottom of anchor expected 85 but got 400
FAIL Should use the third fallback position with enough space above assert_equals: Anchored element should be at the top of anchor expected 235 but got 500
FAIL Should use the second fallback position with enough space right assert_equals: Anchored element should be at the bottom of anchor expected 235 but got 400
FAIL Should use the first fallback position with enough space above and right assert_equals: Anchored element should be at the top of anchor expected 235 but got 500
